ctlra: add feedback_digits() to tcc, improve error handling
This commit adds the new ctlra API to push a number to a set
of 7-segment displays. The driver mushes the provided float
into individual digits to display.

TCC error handling is improved, for when a function call is
made that doesn't have its symbols added to the TCC compile.
This no longer crashes, instead being handled by just ignoring
the script.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>