mappa: POC working with virtual device + crossfade
This commit fixes a bug in the CtlraReader mappa instance,
and registers a target for the crossfader in Mappa. The
control over the crossfader target is mapped by the config
file "mixxx_z1.ini".

As of this commit, we can collaborate on mappings for the Z1,
test them with virtual devices (if you don't have access to
the hardware). This is the milestone that we've worked towards
since December of '16!

Previous PR available here: https://github.com/mixxxdj/mixxx/pull/1308/

Note that this new approach has a significantly different design
from the application side - using "Mappa" to perform the binding
of hardware controls to software targets. Hence, it does not rely
on TCC or JavaScript or ... it just loads a human-readable config
file.

The config file itself has enough power to enable users create complex
and powerful multi-layered mappings - without the application (Mixxx)
or the device (all hardware devices supported by Ctlra) to know about it.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>