mappa: some more controls added
This commit adds volume faders for channels, and in doing so
we create multiple targets with different ranges. The scaling
could be performed at the app level - but that would require
every application to do scaling - and the user wouldn't have
and accurate view of the target's scale. Next steps is to
investigate how we can have the application expose that metadata
to the mappa library.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>