mappa: use mappa for target ranges
This commit uses the latest Ctlra/Mappa commit to enable
range-scaling of each target. This means that Mixxx exposes
the range of each target (eg: crossfader 1 to -1, fader 1 to 0),
and that Mappa will automatically scale a mapping from hardware
to the software range.

In future, users will be able to "range-map" a hardware control
to control not the full software scale, but select a specific
range inside the application provided range. This is useful for
eg lowpass filters, where 0Hz is too low, and users may wish to
customize the "lowest-frequency" that the control-surface can set.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>