control: print [Master] control object creation
This shows how many useless control-proxys there are
in the system - eg: VuMeter_down_small etc. Its not
a good list to present to a user in a UI to map.

We need to
A) Manually code them into Mappa
   (bad - manual work and maintenance)
B) Provide a way to indicate "mappability" at creation time
   (bad - manual work and maintenance)
C) Provide a list of config objects to create targets for
   (good - can be updated by user, no re-compiles required)

Need to see where C) fits best.

Signed-off-by: Harry van Haaren <harryhaaren@gmail.com>